+++ rt-carousel.php
@@ -10,7 +10,7 @@
  * Domain Path: /languages
  * License:     GPL-2.0-or-later
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html
- * Version:     2.0.0
+ * Version:     2.0.1
  * Text Domain: rt-carousel
  *
  * @package rt-carousel